+++ src/common/components/Button/index.styles.tsx
@@ -3,5 +3,5 @@ import styled from 'styled-components';
 import { Button } from '@olist/united';
 
 export const ButtonBorderless = styled(Button)`
-  border: 0px;
+  border: none;
 `;
