+++ src/common/components/Logo/Logo.stories.tsx
